requirements.txt
numpy
matplotlib
pypianoroll>=0.3.0
tensorflow-gpu>=1.4.0